1차 랜더링 개선 useRef
