+++ src/components/TestCompo.tsx
@@ -1,0 +1,6 @@
+function TestCompo() {
+	console.count();
+	return <div>Test Compo</div>;
+}
+
+export default TestCompo;
